Improve QPACK encoding algorithm.
If a header field matches an entry in the dynamic table, it might not be
possible to use it because:
  * the limit on the number of blocked streams is reached, or
  * the entry is considered draining, but cannot be duplicated because there is
    no room in the dynamic table, or
  * the entry is considered draining, but cannot be duplicated because the
    encoder stream buffer limit is reached.

In this case, the entry is currently encoded as a literal header name and
literal header value. This CL encodes the entry instead making use of the name
of an existing static or dynamic table entry, if possible.

Protected by FLAGS_quic_reloadable_flag_quic_better_qpack_compression.

PiperOrigin-RevId: 636163089