Automated g4 rollback of changelist 637261323.
*** Reason for rollback ***

spend_token_data.proto update broke chromium build.

*** Original change description ***

Upgrade to the latest protobuf edition

This change has been verified to be behavior preserving both in the generated code and on the wire.  The only potential changes are for code that uses syntax-dependent reflection.  We've already migrated most of these to safer patterns, and are relying on TAP to catch any uncaught ones.

See go/protobuf-edition-zero-lsc for more information.

#busy-beavers

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:637251876:BASE:6...

***

PiperOrigin-RevId: 637735718