Use human-readable BUG IDs in quic_interval_deque.h
This is a follow-up to David's comment on cl/633296228.

Note that while I believe a BUG ID could have space in it, I did not find a
precedence for that. Most existing IDs are random numeric identifiers,
references to Bugs, or brief phrases formatted with underscore instead of space,
so this CL follows the last option.

PiperOrigin-RevId: 637875266