Make default impl use quiche_feature_flags_list.h instead of quic_flags_list.h.
The two files list the exact same flags, except the name prefix is different
(quiche_ versus quic_), therefore the global variable used to implement the
flags will have a different name. The second value in
quiche_feature_flags_list.h is the same as the only value in quic_flags_list.h.
This CL should therefore not result in any behavior change.

The default flag implementation is only used in Chromium, and I already prepared
a CL necessary to roll this QUICHE change.

PiperOrigin-RevId: 647660906