Update PreSharedKey tests to pass without non-fatal QUIC_BUG
The PreSharedKey tests were relying on a QUIC_BUG call to not terminate program execution, but this seems like an implementation detail (one that might be worth changing in the Chromium implementation).

It seems that there was some related work on this in https://github.com/google/quiche/commit/b45e7741802e3b1fbe5635ea6798df0ce0991a16, but the change there isn't enough to get the test passing locally for me for some reason. 🤔

PiperOrigin-RevId: 650760186