Refactor MoQT sending logic to pull objects from the application, instead of pushing them onto individual MoQT sessions.
This CL introduces a new interface, a MoqtTrackPublisher.  A track publisher knows everything necessary to publish a track, caches recent objects and may be shared across multiple MoQT sessions.  Objects are always pulled from the publisher.

This CL also restructures the sending logic around every object being published within a single subscription (there has been a lot of code from the era where we deduplicated objects; that is now gone).  The publisher notifies the subscription, which in turn notifies the relevant stream within the session.

There is also a lot of other minor refactors as necessary.

The current version of this CL does not support publishing objects far in the past, largely due to design ambiguities surrounding how fetch is supposed to work (I am writing a slide deck for Vancouver on that).  Instead, we merely backfill the recent objects that are still in the cache; that covers most of the use cases we care about for now.

PiperOrigin-RevId: 653833971