Allocate memory less often in HttpHeaderBlock hash function
quiche::StringPieceCaseHash used a std::string to store the lower-cased version of the string to be hashed. This permitted 22 character header names to be hashed without memory allocation on 64-bit platforms, but the limit is much less on 32-bit platforms, which affects Chrome.

Instead, use an absl::FixedArray with internal storage sized for a cache line to store the temporary string. This should be enough to avoid heap allocation for any real-world header name.

Also add a unit test.

Not flag-protected as it is a low-risk optimization with no functional change.

PiperOrigin-RevId: 660691522