Factor out size limit logic from QuicHeaderList.
QuicHeaderList is used both in IETF QUIC and in gQUIC. The size limit
functionality is only used in gQUIC. (In IETF QUIC, size is limited in
QpackDecodedHeadersAccumulator). This CL creates a wrapper class for use in
gQUIC only. When gQUIC gets deprecated, this wrapper class can be cleanly
removed, together with QuicSpdySession::SpdyFramerVisitor.

PiperOrigin-RevId: 662112283