Add ability to run QuicDispatcher without any versions
When I added this QUIC_BUG back in cl/284274377, my assumption was that we'd never want to run a QuicDispatcher with empty versions. It turns out that this assumption was wrong: it's useful to do this if we want the dispatcher to still have the ability to send version negotiation packets to avoid client timeouts. This is a feature we need for Privacy Proxy, where we want Chrome to fall back quickly if the proxies are unavailable. In particular, when we disable IP Protection, we want to put the dispatcher in this state.

I'll also note that this QUIC_BUG wasn't particularly useful, since it only checked at construction time. It was possible to construct a dispatcher with a version and then use reloadable flags to remove that version, causing the dispatcher to land in this state without firing any QUIC_BUGs.

So this CL removes the QUIC_BUG, and adds test to confirm that the dispatcher behaves correctly in this state - which is to respond to packets from all versions with version negotiation packets.

This change is not flag-protected, because it is never reached in production today (if it were, the QUIC_BUG would be firing).

PiperOrigin-RevId: 662723304