Update QuicStream to handle RESET_STREAM_AT frames.
When received, the stream will make a note of the reliable_size and store a copy of the RESET_STREAM_AT.

If the stream has already consumed at least reliable_size bytes, processes the copy as if it were a regular RST_STREAM.

If not already consumed, QuicStream checks every time data is consumed to see if the reliable size has been exceeded, and when it is, processes the reset unless the FIN bit has already been consumed.

This CL is just the QuicStream. Other components will follow in later CLs.

Protected by FLAGS_quic_reloadable_flag_quic_reliable_stream_reset.

PiperOrigin-RevId: 675988866