Add a new OnParsedClientHelloReceived method to QuicConnection and QuicConnectionDebugVisitor.
This method will be called after the QuicSession is created and the ClientHello is parsed by the server. The QuicConnection will simply forward the parsed CHLO to the debug visitor. The base debug visitor implementation will do nothing.

I plan to use this in BDN to log the cert compression algorithms, and potentially other fields of the parsed CHLO.

PiperOrigin-RevId: 681963634