Implement QuicAlarmMultiplexer.
The alarm multiplexer allows all alarms for the connection to use the same underlying platform alarm, and to defer updates for that alarm until all connection alarms are processed.  This reduces the number of update calls to the underlying platform alarm, which can be fairly expensive.

Protected by FLAGS_quic_reloadable_flag_quic_use_alarm_multiplexer.

PiperOrigin-RevId: 690982683