Change protection scheme for code that marks outgoing ECT.
The code that marks outgoing packets ECT(0) or ECT(1) was originally written to support tests of the code that receives and reports those marks, although we knew it would eventually be used to send them in production as well. The plan was to eventually experiment with sending ECT(1) from GFE, so most of the code was protected with quic_support_ect1.

The plan has changed. For multiple reasons, initial send-side experiments will occur in Chrome instead. As such, a GFE feature flag is a cumbersome means to manage a rollout.

Instead, sending ECT marks will only be enabled when QUIC loads a congestion controller that supports it. There is no code path in Quiche or GFE that will do this. Any addition in this space can be flag-protected in case of unexpected bad effects.

The various QuicPacketWriter derived classes now simply report 'true' for ECN support since they are no longer dependent on the flag. This is necessary but not sufficient to send ECT marks, because the congestion controller also needs to return true for 'EnableECT1' or 'EnableECT0'.

Also, fix inconsistent checking of the ECN setting before calling set_ecn_bits().

This CL dramatically reduces the footprint of the feature flag. Most new code is trivial when the connection is not sending ECT. There are two exceptions:

1) Feature flag quic_preserve_dscp_with_ecn protects the code that reads the DSCP bits currently set from a socket and ORs them with whatever ECN setting it is propagating. This code is a bit more complex and therefore higher-risk. It is not used in Chrome, and I strongly suspect it is not used in other Quiche codepaths. (Only QuicDefaultPacketWriter sets the ECN info bit, and that class is only used or extended in fringy places).

2) QuicSentPacketManager has some complex ECN validation code in its ACK processing. This is protected by a new boolean member ecn_queried_ so that it only executes if the QuicConnection has turned on ECN markings.

Protected by FLAGS_quic_restart_flag_quic_preserve_dscp_with_ecn.

PiperOrigin-RevId: 695874210