Some more QUICHE build fixes
These were issues found by trying to get quiche building with the latest Bazel. The latest Bazel seems to do more include layering checks.

This isn't sufficient to fix the build with Bazel 7 (there is a thornier issue with how quiche_platform_impl is wired up, and also some weird interaction with bzlmod), but gets part of the way there.

PiperOrigin-RevId: 696213724