Update ECN tests to function properly in Guitar.
cl/695874210, while deprecating the previous flag protection for send-side ECN, identified a particular block of code as relatively high-risk and so protected it with flag quic_preserve_dscp_with_ecn. Unfortunately, the code that executes when this flag is false does not correctly set the ECN bits when a test is running on Guitar (which uses GOOGLE_IP_TOS), causing failures.

This CL deletes the flag, and protects the entire code block by only executing it when packet_info.ecn_codepoint is not NOT_ECT (which never happens in production today).

Tests now pass in Guitar.

PiperOrigin-RevId: 696290469