Refactor QuicChaosProtector
This CL refactors the chaos protector to have it ingest an array of QuicFrames instead of a single QuicCryptoFrame and padding. This CL does not change behavior for existing code, it prepares it for a future CL that will add multi-packet chaos protection.

Portions of this change were written by wub@, in particular:
- Added code to prevent reordering of ACK frames, to minimize impact on congestion control.
- Added unit tests for the cases when input frames are `[crypto, crypto, padding]` or `[ack, crypto, padding]`.

Protected by FLAGS_quic_reloadable_flag_quic_enable_new_chaos_protector.

PiperOrigin-RevId: 696443185