Add oghttp2 and nghttp2 testing to demonstrate client GOAWAY-sending behavior.
This CL adds analogous testing to OgHttp2AdapterTest and NgHttp2AdapterTest to
exercise the scenario where a client sends a request/receives a response and
then submits a GOAWAY based on the adapter GetHighestReceivedStreamId().

For oghttp2, GetHighestReceivedStreamId() is updated to the last stream ID that
was processed in some way, i.e., the stream ID of the request/response stream.
For nghttp2, GetHighestReceivedStreamId() is kept at 0 because the stream was
not initiated by the peer (server push). Currently, the submitted GOAWAY is
sent from both adapters, but with these different last-stream-ID fields.

The future plan will be to have oghttp2 avoid sending GOAWAYs as clients
altogether, as sending a GOAWAY as client is unintuitive.

PiperOrigin-RevId: 697766519