Simplify how quiche platform files are incorporated
There's a circular dependency between, say, quiche_core and
quiche_platform_default. They're separated to apply strip_include_prefix, but
this makes newer Bazel flag the unmarked dependency. The export headers were
special-cased to work around this, but now logging headers are also part of
this. If we instead just use includes, they can be part of the same target
(which they logically are) and the circular dependency issue is gone.

This fixes another blocker to updating QUICHE to newer Bazel.

PiperOrigin-RevId: 698514759