Remove ResumptionAttempted assertion in TlsServerHandshakerTest.ResumptionWithPlaceholderTicket
This is a bit of a mess. QUICHE started sending a placeholder ticket when ticket encryption failed, because it used to return an empty string, which was not actually valid. However, the flag was not consistently enabled, so sometimes the test ran with the old empty string.

BoringSSL happened not to notice, but newer BoringSSL clients catch this. However, newer BoringSSL servers, as of https://boringssl.googlesource.com/boringssl/+/c59bf8bf189dcbde868e04efcd53b705ed155231, now also support empty string as a signal to skip sending a ticket. That is actually the ideal behavior for QUICHE in this situation.

However, the tests actually reject the ideal behavior because they assert that `ResumptionAttempted()` is true. Remove the assertion so that QUICHE, with the flag on or off, is compatible with the newest BoringSSL.

PiperOrigin-RevId: 698517697