Group mock_load_balancer_encoder.h with load_balancer in the standalone build
By default it ends up in `quiche_test_support`, but it depends on a header in `load_balancer` that isn't built in the standalone build. Talking to vasilvv, it sounds like the preferred pattern is to, like the moqt section, put it with `load_balancer` so it gets out of the way.

This fixes a build failure with Bazel 7, which is a bit stricter about headers.

PiperOrigin-RevId: 698774080