Unblock quic_enable_chaos_protection_v2 flag
Now that Chaos Protection v2 is feature complete and has good coverage in both unit tests and end-to-end tests, we're ready to start deploying it.

This flag was originally called --gfe2_reloadable_flag_quic_enable_new_chaos_protector but multiple CLs landed behind it so we need to rename it before unblocking. We're renaming it to --gfe2_reloadable_flag_quic_enable_chaos_protection_v2.

This CL unblocks --gfe2_reloadable_flag_quic_enable_chaos_protection_v2 for deployment on servers, and enables it in Chrome.

PiperOrigin-RevId: 704858299