Fix flakiness in chaos protection test
This test would sometimes fail because chaos protection isn't guaranteed to always create this many CRYPTO, PING or PADDING frames if it runs out of available space. To resolve this, the test now checks if any of the symptoms of chaos protection is present, which is enough to ascertain that chaos protection was indeed performed.

Prior to this change, any run of this test with --runs_per_test=25 would reproduce the failure. With this fix, testing with --runs_per_test=1000 passes.

This CL also fixes a spelling mistake in a debug-only chaos protection log.

PiperOrigin-RevId: 707138040