Switch QUIC version negotiation to RFC codepoint
This CL replaces the codepoint from draft-ietf-quic-version-negotiation-13 with the one from RFC 9368. The flag protection is designed so that we'll be fully switched over once we deprecate the quic_version_negotiation_rfc flag.

It's worth noting that this means that older software won't interoperate with new software when it comes to this extension. This is acceptable because this extension is currently not relied upon: since we don't support compatible version negotiation, the only benefit of this extension is to prevent version downgrades, but right now our code only supports versions of IETF QUIC with equal security properties, so it does not actually provide any protections at the moment.

Note that this also means that the strings and logs don't differentiate between the old and new value, so logs will be slightly ambiguous during the transition period if we get a log without knowing the value of the flag.

Alternatively, we could have instead used a succession of flags with support for multiple versions throughout a lengthy transition, but that would have significantly increased the engineering work and delayed the transition for no observable benefit.

Protected by FLAGS_quic_reloadable_flag_quic_version_negotiation_rfc.

PiperOrigin-RevId: 708402132