Officially allow out-of-order writes in QuicStreamSendBuffer
This has always worked, but there was a QUIC_BUG to prevent it. Unfortunately the QUIC_BUG had a bug where QuicStreamSendBuffer::current_end_offset_ was not computed properly, so it never correctly enforced in-order writes. Chaos Protection v2 triggers this QUIC_BUG when the Client Hello size exceeds the value of the quic_send_buffer_max_data_slice_size flag.

PiperOrigin-RevId: 712548578