Update ParseQuicTag to only use hex decoding if valid
This relands cl/616801281. It should be safe to land now that Envoy uses a newer revision of abseil (see https://github.com/envoyproxy/envoy/pull/36317)

ParseQuicTag was using absl::HexStringToBytes to unconditionally decode strings when they were 8 bytes long, but for strings with non-hexidecimal characters the return value of calling absl::HexStringToBytes is unspecified. This CL updates ParseQuicTag to only use the output of absl::HexStringToBytes when the call returns success.

This also moves the code away from using the deprecated version of absl::HexStringToBytes, allowing the Chrome build of QUICHE to build without the `-Wno-deprecated-declarations` clang flag.

PiperOrigin-RevId: 714809406