Delete the code that sends the MinAckDelay transport parameter from draft-iyengar-quic-ack-frequency-01. Receiving the parameter is still supported.
This has been replaced with a new value (and new frame format) in later drafts. For now, we need to maintain the ability to process the parameter from old clients, but no up-to-date client should send it.

The connection option that controlled this (AFFE) has been retired from any active experiments.

PiperOrigin-RevId: 719396126