Switch QUICHE to bzlmod
In doing so, update various Bazel dependencies. Ideally we would update
com_google_googleurl to the bzlmod-aware version, but I'm not sure how
uploading the storage buckets works, and it seems to work fine as-is.

This required changing the zlib.h include to use angle brackets because
otherwise the include checker, with BCR's zlib package, fails.

PiperOrigin-RevId: 731365105