Make MoqtControlParser read directly from the stream, similar to what MoqtDataParser already does.
The immediate cause for this is a recurring (but very difficult to isolate bug) in which something like this happens:
1. Parser receives a control frame.
2. The application visitor does not like the contents of the frame, so it terminates the connection.
3. MoqtSession closes the underlying WebTransport session, which results in the control stream being reset.
4. Control stream being reset results in the receive buffer being deleted.
5. When the visitor returns, the parser attempts to parse the next chunk, not realizing that the buffer in question has been deleted.

I spent some time trying to come up with a workaround, but none of them seemed to work, so I gave up and just rewrote the thing to call the stream directly; that way, if the stream drops its buffers, the read will just fail.

(this CL does not have a repro for this bug, since triggering it consistently turns out to be really hard due to the way parser does its own buffering; a downstream CL has an ASAN crash that triggers it, and I've also randomly hit it in the integration tests before)

As a side effect, I also made the parser automatically fail on any FIN on the control stream (we used to handle those at message boundaries, which is not necessary).

PiperOrigin-RevId: 744787214