Comment out absl nullability macros when copying out from google3/.
In some of the versions of absl that quiche is used with, the macros are uppercase; in others, the macros are lowercase, so until this is consistent, comment them out. This workaround will allow go/nullability-annotation-alias-to-macro-migration to proceed, without significant loss of verification of the annotations, which is done by an internal-only ClangTidy check in google3.

This has no impact on code internally, and fixes the external build.

PiperOrigin-RevId: 745256376