Allow connect-udp to work across connection ID changes
Prior to this CL, the MASQUE backend code would use the QUIC connection ID as a key into a hashtable. Back when this was written, the CID of a QUIC connection was stable for the lifetime of the connection, so this worked. Since then, we've now added support for evolving CIDs, so this code no longer worked when interoperating with implementations that aggressively retire CIDs. This CL fixes that by using a vector instead of a hashtable, and comparing pointers instead of performing a CID lookup.

PiperOrigin-RevId: 770297284