Add new static QUIC tag for explicit connection close packet
This CL adds new static QUIC tag, ECCP, to explicitly send connection close packet when server closes the connection with silent idle timeout. This option will be sent from the Chromium to ask the server to send an explicit connection close packet when the server decides to close the connection silently with silent idle timeout.

PiperOrigin-RevId: 770668307