Consume the data buffer on early return in `OnStreamFrameData`
This commit calls the `Consume` method on the stream ID when an early
return is encountered in the `OnStreamFrameData` method. The rational is
that the data is marked in the first instruction of the method and
should be released even when an error is encountered. The impact of not
releasing the marked data is that the connection-level window manager
may not send `WINDOW_UPDATE` frames blocking the sender forever.