Stop sending Host header in MASQUE HTTP/2 requests
In HTTP/2, the `:authority` header is mandatory and the `host` header is optional. The only requirement is that, if `host` is included, it needs to match what's in `:authority`. So this code was doing something allowed by [the spec](https://www.rfc-editor.org/rfc/rfc9113#section-8.3.1-2.3.1), but some servers fail to process it. Since this is just a test client, we'll work around that issue by only sending `:authority`.

PiperOrigin-RevId: 782086853