Update Subgroup Stream Header formats to draft-11.
Also eliminates "End of Track and Group" object status.

Subgroup ID is no longer optional in Object Metadata. For Datagrams, it is equal to the object ID.

SerializeObjectHeader() has two versions, currently: one that is compact and easy to understand, but involves a little more copying, and another that is the opposite.

PiperOrigin-RevId: 783016191