Simplify the way bitrate adjuster provides bitrate feedback.
We used to have logic to throttle how often one can change bitrate, but apparently this is unnecessary in a lot of encoders, since they can adjust bitrate without cutting a new keyframe.  Instead, always notify the encoder and let it decide itself whether to adjust or not.

Also move parameters for bitrate adjuster from kConstants to a parameter struct.

PiperOrigin-RevId: 785547333