Update MOQT FETCH to draft-12.
Mostly, fields are just renamed, but the encoding of FETCH_OK End Location did change. Also changed some internal FETCH APIs to avoid an optional for end_object, instead just setting it to kMaxObjectId if the subscriber wants the whole group.

Also the interface between Joining Fetch processing and the Publisher was fatally broken. Joining Fetch must be forwarded upstream as-is, not with a priori knowledge of Largest Location.

PiperOrigin-RevId: 791399297