Update MoQT Stream Header subgroup to draft-12.
The type can now encode that there is an end of group in the stream.

There are deep flaws in the specification for this, so this CL is close to the minimum diff to maintain interoperability until the design changes substantially. As a result, there are no new tests for this functionality and most existing tests use the non-end-of-group stream types.

PiperOrigin-RevId: 792708909