Add 2 new classes QuicConnectionMigrationManager and QuicSpdyClientSessionWithMigration which are extracted from Chromium [QuicChromiumClientSession](https://source.chromium.org/chromium/chromium/src/+/main:net/quic/quic_chromium_client_session.h) to handle network disconnected signal from the platform by migrating to an alternative network if available.
Note that identifying the platform picked default network and migrating back to default network after the above migration will be implemented as a follow up.

Added a factory interface QuicPathContextFactory for QuicPathValidationContext.

This is part of the chain CLs to fork Chromium's connection migration code extracted in https://chromium-review.googlesource.com/c/chromium/src/+/5980615 which is verified to work for Chromium.

PiperOrigin-RevId: 794210361