Fix MigrationStatus histogram calls in quic_connection_migration_manager.cc
quic_connection_migration_manager.cc defined enums with the maximum value set to MIGRATION_STATUS_MAX. However, since QuicConnectionMigrationStatus is an enum class, the value needs to be qualified with the enum name.

This was causing the Chromium build to fail.

Fix it.

No functional change in the GFE (the modified argument is unused outside Chromium).

PiperOrigin-RevId: 796398774