Support various migrations upon detecting path degrading: Perform early migration to a different network if allowed. Otherwise migrate to a different port if allowed. Otherwise do nothing.
Also add logic to handle platform signal OnNewNetwork(). If there was an earlier attempt to migrate to a different network but it failed due to no alternative network available, try migrate again.

This is part of the chain CLs to fork Chromium's connection migration code extracted in https://chromium-review.googlesource.com/c/chromium/src/+/5980615 which is verified to work for Chromium.

PiperOrigin-RevId: 801032706