Have BlindSignAuth own its dependency to BlindSignMessageInterface (1/2).
This CL adds an additional constructor for BlindSignAuth, which takes in a unique_ptr instance and owns such pointer. We are moving to using this constructor for BlindSignAuth. There's a dependency in IpProtectionTokenDirectFetcher that must use the old constructor implementation, once that's refactored in chromium, we can go ahead and delete the old constructor.

PiperOrigin-RevId: 805932956