Add TaskBundle abstraction to fan out heavyweight ops
Generating blinded tokens and unblinding tokens are both fairly CPU-intensive, so it could be beneficial to perform those operations in parallel when operating on batches of tokens. By providing alternate implementations of TaskBundle, users can inject their preferred executor (e.g. by posting to a TaskRunner in Chromium, or using a pool of Fibers in a Scaffolding server).

PiperOrigin-RevId: 810903802