Add fuzzer for QuicheDataWriter
This CL adds a new FuzzTest-based fuzzer that exercises sequences of method calls on a QuicheDataWriter object.

This goes further than the typical "DoesNotCrash" property by verifying that the corresponding operations in QuicheDataReader have the expected results. (For instance, if you write a uint8_t, reading a uint8_t should return the same value.)

PiperOrigin-RevId: 810967573