Delay closing QUIC connections on invalid SNI and log the normalized SNI.
This change defers closing the connection due to an invalid SNI until after the transport parameters have been processed so that the SNI received in ClientHello can be compared with the SNI received in the transport parameter. This allows for logging the normalized SNI from the ClientHello, even in cases where the SNI is invalid.

Protected by FLAGS_quic_reloadable_flag_quic_delay_connection_close_on_invalid_sni.

PiperOrigin-RevId: 815464154