Add fuzz tests for SNI normalization and validity properties in QuicheHostnameUtils
Primarily, this tests that NormalizeHostname() is idempotent for hostnames where IsValid() returns true. It also tests that normalizing a valid hostname produces another valid hostname.

If normalization is not idempotent, we'd have to think carefully about whether the server ever normalizes incoming SNIs twice.

If normalization does not preserve validity, we'd have to make sure the client checks validity after normalization. (Otherwise, a client could send what they believe is a valid SNI, only for the server to reject it post-normalization.)

PiperOrigin-RevId: 823669504