Detect and close connections on invalid acks received by the QUIC dispatcher.
With this change, the QUIC dispatcher will track the largest packet number among the packets it has sent. When processing incoming packets, the `TlsChloExtractor` will check if any ACK frames in the received packet, acknowledge packet numbers that were not sent by the dispatcher. If such an "invalid ack" is found, the connection is considered to be in violation of the protocol and is statelessly closed by adding the connection ID to the time-wait list.

Protected by quic_restart_flag_quic_dispatcher_close_connection_on_invalid_ack.

PiperOrigin-RevId: 826482880