Refactoring Indeterminate-Length BHTTP message section and tests.
This change refactors the indeterminate-length BHTTP request decoder to use helper functions that will be needed for implementing the encoder. This also moves the private message section enums so it can be accessed by the to-be-implemented encoder/decoder and refactors the tests to re-use the encoding for the future encoder tests.

PiperOrigin-RevId: 833500305