Make it a `QUIC_BUG` to access the `QuicSession`'s `QuicConfig` after the handshake completes, in preparation for dropping the config completely to safe memory.
This CL moves the `QuicConfig` object into a new `SavedConfig` class
which the session stores instead. This new class provides accessors
to the underlying config which trigger the `QUIC_BUG` once the config
has been "deleted". (The actual deletion will happen in a follow up CL).

It also provide a new accessor for access to the config after deletion,
which callers will need to use to avoid the `QUIC_BUG`. In a follow-up
(flag-protected) CL, the `SavedConfig` will cache values from the
`QuicConfig` when it is deleted, and return those instead.

Project import generated by Copybara.

PiperOrigin-RevId: 836879510