Improve error handling in OHTTP prober
The main issue here was that the OhttpProbeClient::callback_response_status_ member variable was never set anywhere, so there was no way for it to not be OK.

While I was in there, I also caught a few other minor failure modes where the status wasn't set properly. I also more added validation of the response.

PiperOrigin-RevId: 841205501