Propagate forwarding preference across the MOQT relay.
Currently, objects that arrive as datagrams go out in subgroups. This adds forwarding preference to object metadata so that it is reported in OnObjectFragment().

This CL is written in the knowledge that forwarding preference is change from a a track property to an object property in a later draft. So there are no new checks for consistency across the track.

PiperOrigin-RevId: 841855517