Speed up the slow e2e test TestInvalidAckBeforeHandshakeClosesConnection
It turns out that this test was trying to make a connection after setting packet loss to 100%, which meant we waited the full 30-second timeout.

Now, we just let it create a dummy connection so we can steal its writer, which we use to write the invalid ACK frame. The test succeeds in 5 seconds on my machine instead of >30.

PiperOrigin-RevId: 843859547