Properly handle query parameters in MASQUE/OHTTP test code
In HTTP/2 and HTTP/3, the ":path" pseudo-header is poorly named in that it carries not only the path but also the query parameters. That confusion led to bugs in this codebase where only the path was sent or acted upon. This CL fixes that by sending both and making sure to ignore the query upon receipt when relevant.

PiperOrigin-RevId: 847141777