QBONE ICMP reachable: convert to "ping sockets" to avoid the need for `SOCK_RAW` permissions.
["ping sockets"](https://lore.kernel.org/all/20110510180957.GA3262@albatros/) are a kernel feature from ~2011 that allow initiating ICMP echo requests via `SOCK_DGRAM` with protocol `IPPPROTO_ICMPV6`, instead of requiring `SOCK_RAW`.

Once such a socket is created, the kernel expects `send`/`sendto` payloads to start directly with the ICMP header.

The kernel commit message says, in part:
>ICMP headers given to send() are checked and sanitized.
>
> (... ) The id is set to the number (local port) of the socket

which could perhaps have been written more clearly as "The id MUST BE set to the local port", as in my experience not doing so results in `EINVAL`.

Because QBONE may ping addresses that are not explicitly listed in the routing table, additionally force the destination address scope ID to be that of the TUN/TAP interface used by the bonnet.  For typical usage (tun device `qbone0`) this is the equivalent of the `%qbone0` bit of the command line:
```
$ ping -6 <destaddr>%qbone0
```

PiperOrigin-RevId: 847850750